# PERF: MultiIndex.memory_usage shouldn't trigger the index engine