--- a/pandas/tests/base/test_misc.py
+++ b/pandas/tests/base/test_misc.py
@@ -142,31 +142,6 @@ def test_memory_usage_components_narrow_series(any_real_numpy_dtype):
     assert total_usage == non_index_usage + index_usage
 
 
-def test_memory_usage_doesnt_trigger_engine(index):
-    index._cache.clear()
-    assert "_engine" not in index._cache
-
-    res_without_engine = index.memory_usage()
-    assert "_engine" not in index._cache
-
-    # explicitly load and cache the engine
-    _ = index._engine
-    assert "_engine" in index._cache
-
-    res_with_engine = index.memory_usage()
-
-    # the engine doesn't affect the result even when initialized with values, because
-    # index._engine.sizeof() doesn't consider the content of index._engine.values
-    assert res_with_engine == res_without_engine
-
-    if len(index) == 0:
-        assert res_without_engine == 0
-        assert res_with_engine == 0
-    else:
-        assert res_without_engine > 0
-        assert res_with_engine > 0
-
-
 def test_searchsorted(request, index_or_series_obj):
     # numpy.searchsorted calls obj.searchsorted under the hood.
     # See gh-12238